refactor: rename and remove commented jsx
